LayoutTests/fast/css-grid-layout/grid-columns-rows-get-set-expected.txt

Test that setting and getting grid-columns and grid-rows works as expected

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Test getting -webkit-grid-columns and -webkit-grid-rows set through CSS
PASS getComputedStyle(gridWithNoneElement, '').getPropertyValue('-webkit-grid-columns') is 'none'
PASS getComputedStyle(gridWithNoneElement, '').getPropertyValue('-webkit-grid-rows') is 'none'
PASS getComputedStyle(gridWithFixedElement, '').getPropertyValue('-webkit-grid-columns') is '10px'
PASS getComputedStyle(gridWithFixedElement, '').getPropertyValue('-webkit-grid-rows') is '15px'
PASS getComputedStyle(gridWithPercentElement, '').getPropertyValue('-webkit-grid-columns') is '53%'
PASS getComputedStyle(gridWithPercentElement, '').getPropertyValue('-webkit-grid-rows') is '27%'
PASS getComputedStyle(gridWithAutoElement, '').getPropertyValue('-webkit-grid-columns') is 'auto'
PASS getComputedStyle(gridWithAutoElement, '').getPropertyValue('-webkit-grid-rows') is 'auto'
PASS getComputedStyle(gridWithEMElement, '').getPropertyValue('-webkit-grid-columns') is '100px'
PASS getComputedStyle(gridWithEMElement, '').getPropertyValue('-webkit-grid-rows') is '150px'
PASS getComputedStyle(gridWithViewPortPercentageElement, '').getPropertyValue('-webkit-grid-columns') is '64px'
PASS getComputedStyle(gridWithViewPortPercentageElement, '').getPropertyValue('-webkit-grid-rows') is '60px'
Test getting wrong values for -webkit-grid-columns and -webkit-grid-rows through CSS (they should resolve to the default: 'none')
PASS getComputedStyle(gridWithFitContentElement, '').getPropertyValue('-webkit-grid-columns') is 'none'
PASS getComputedStyle(gridWithFitContentElement, '').getPropertyValue('-webkit-grid-rows') is 'none'
PASS getComputedStyle(gridWithFitAvailableElement, '').getPropertyValue('-webkit-grid-columns') is 'none'
PASS getComputedStyle(gridWithFitAvailableElement, '').getPropertyValue('-webkit-grid-rows') is 'none'

Test the initial value
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-columns') is 'none'
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-rows') is 'none'

Test getting and setting -webkit-grid-columns and -webkit-grid-rows through JS
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-columns') is '18px'
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-rows') is '66px'
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-columns') is '55%'
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-rows') is '40%'
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-columns') is 'auto'
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-rows') is 'auto'
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-columns') is '80px'
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-rows') is '150px'

Test setting grid-columns and grid-rows back to 'none' through JS
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-columns') is '18px'
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-rows') is '66px'
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-columns') is 'none'
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-rows') is 'none'
PASS successfullyParsed is true

TEST COMPLETE
